requirements.txt
openai
python-dotenv
#dotenv
streamlit
pandas
qdrant-client
#PIL
#base64